+++ internal/acceptance_tests/blocks/fastly_service_authorization.tf
@@ -1,0 +1,5 @@
+resource "fastly_service_authorization" "test" {
+  service_id = {{.SERVICE_ID_REF}}
+  user_id    = "{{.USER_ID}}"
+  permission = "{{.PERMISSION}}"
+}